Datasets support + minor fixes

+++ bench/main-binary.cpp
@@ -8,11 +8,11 @@ using namespace std;
 
 uint8_t* generateSequences(BenchmarkParams &bParams, ExperimentParams &xParams) {
     if (bParams.verbose) cout << "Generation of data..." << std::endl;
-    xParams.bytesPerSequence = ceilDivisionBySmallInteger(xParams.m, xParams.bitsPerPacked)
+    xParams.bytesPerSequence = (int) ceilDivisionBySmallInteger(xParams.m, xParams.bitsPerPacked)
             * ceilDivisionBySmallInteger(xParams.bitsPerPacked, 8);
     if (xParams.alignSequences)
-        xParams.bytesPerSequence = ceilDivisionBySmallInteger(xParams.bytesPerSequence, 8) * 8;
-    uint8_t* sequences = new uint8_t[xParams.d * xParams.bytesPerSequence]();
+        xParams.bytesPerSequence = (int) ceilDivisionBySmallInteger(xParams.bytesPerSequence, 8) * 8;
+    uint8_t* sequences = new uint8_t[(size_t) xParams.d * xParams.bytesPerSequence]();
     getRandomValues(sequences, xParams);
     return sequences;
 }
